+++ nacc_form_validator/errors.py
@@ -36,13 +36,12 @@ class ErrorDefs:
     CHECK_GDS_2 = ErrorDefinition(0x2005, "compute_gds")
     CHECK_GDS_3 = ErrorDefinition(0x2006, "compute_gds")
     CHECK_GDS_4 = ErrorDefinition(0x2007, "compute_gds")
-    CHECK_GDS_5 = ErrorDefinition(0x2008, "compute_gds")
-    COMPARE_WITH = ErrorDefinition(0x2009, "compare_with")
-    COMPARE_WITH_PREV = ErrorDefinition(0x3000, "compare_with")
-    RXNORM = ErrorDefinition(0x3001, "check_with")
-    DATE_CONVERSION = ErrorDefinition(0x3002, 'compare_age')
-    COMPARE_AGE = ErrorDefinition(0x3003, 'compare_age')
-    COMPARE_AGE_INVALID_COMPARISON = ErrorDefinition(0x3004, 'compare_age')
+    COMPARE_WITH = ErrorDefinition(0x2008, "compare_with")
+    COMPARE_WITH_PREV = ErrorDefinition(0x2009, "compare_with")
+    RXNORM = ErrorDefinition(0x3000, "check_with")
+    DATE_CONVERSION = ErrorDefinition(0x3001, 'compare_age')
+    COMPARE_AGE = ErrorDefinition(0x3002, 'compare_age')
+    COMPARE_AGE_INVALID_COMPARISON = ErrorDefinition(0x3003, 'compare_age')
 
 
 class CustomErrorHandler(BasicErrorHandler):
@@ -103,24 +102,22 @@ class CustomErrorHandler(BasicErrorHandler):
             "If GDS not attempted (nogds=1), there cannot be >=12 questions with "
             + "valid scores - GDS rule no: {0}",
             0x2006:
-            "incorrect total GDS score {1}, expected value {2} - GDS rule no: {0}",
+            "incorrect GDS score {1}, expected value {2} - GDS rule no: {0}",
             0x2007:
-            "incorrect partial GDS score {1}, expected value {2} - GDS rule no: {0}",
-            0x2008:
             "If GDS attempted (nogds=blank), at least 12 questions need to have "
             + "valid scores - GDS rule no: {0}",
-            0x2009:
+            0x2008:
             "input value doesn't satisfy the condition {0}",
-            0x3000:
+            0x2009:
             "failed to retrieve record for previous visit, cannot proceed with "
             + "validation {0}",
-            0x3001:
+            0x3000:
             "Drug ID {0} is not a valid RXCUI",
-            0x3002:
+            0x3001:
             "failed to convert value to a date: {0}",
-            0x3003:
+            0x3002:
             "input value {0} doesn't satisfy the condition: {1}",
-            0x3004:
+            0x3003:
             "input value {0} is not a valid age to compare to {1}: {2}"
         }